Refactor to `InteractsWithSourceFiles` Concern
